requirements.txt
aiofiles           24.1.0
annotated-types    0.7.0
anyio              4.9.0
certifi            2025.6.15
charset-normalizer 3.4.2
click              8.1.8
colorama           0.4.6
fastapi            0.115.14
ffmpeg-python      0.2.0
filelock           3.18.0
fsspec             2025.5.1
future             1.0.0
gTTS               2.5.4
h11                0.16.0
idna               3.10
Jinja2             3.1.6
llvmlite           0.44.0
MarkupSafe         3.0.2
more-itertools     10.7.0
mpmath             1.3.0
networkx           3.5
numba              0.61.2
numpy              2.2.6
openai-whisper     20250625
pip                25.1.1
pydantic           2.11.7
pydantic_core      2.33.2
python-multipart   0.0.20
regex              2024.11.6
requests           2.32.4
setuptools         80.9.0
sniffio            1.3.1
starlette          0.46.2
sympy              1.14.0
tiktoken           0.9.0
torch              2.7.1
tqdm               4.67.1
typing_extensions  4.14.0
typing-inspection  0.4.1
urllib3            2.5.0
uvicorn            0.34.3